+++ packages/wallet-specs/src/protocols/conclude-channel/protocol.ts
@@ -1,4 +1,4 @@
-import { store } from '../..';
+import { store } from '../../temp-store';
 import * as LedgerDefunding from '../ledger-defunding/protocol';
 import * as VirtualDefundingAsHub from '../virtual-defunding-as-hub/protocol';
 import * as VirtualDefundingAsLeaf from '../virtual-defunding-as-leaf/protocol';
