+++ packages/wallet-specs/src/protocols/partial-withdrawal/protocol.ts
@@ -1,17 +1,9 @@
-import {
-  add,
-  Channel,
-  store,
-  subtract,
-  ethAllocationOutcome,
-  checkThat,
-  getEthAllocation,
-} from '../..';
+import { add, Channel, subtract, ethAllocationOutcome, checkThat, getEthAllocation } from '../..';
 import * as ConcludeChannel from '../conclude-channel/protocol';
 import * as CreateNullChannel from '../create-null-channel/protocol';
 import * as LedgerUpdate from '../ledger-update/protocol';
 import { isAllocationOutcome, Allocation, Outcome } from '@statechannels/nitro-protocol';
-
+import { store } from '../../temp-store';
 const PROTOCOL = 'partial-withdrawal';
 const success = { type: 'final' };
