+++ packages/wallet-specs/src/protocols/virtual-defunding-as-leaf/protocol.ts
@@ -1,7 +1,7 @@
 import { assign } from 'xstate';
-import { store, Without, ethAllocationOutcome, checkThat, getEthAllocation } from '../../';
+import { Without, ethAllocationOutcome, checkThat, getEthAllocation } from '../../';
 import { isIndirectFunding, isVirtualFunding } from '../../ChannelStoreEntry';
-
+import { store } from '../../temp-store';
 import * as LedgerUpdate from '../ledger-update/protocol';
 import { isAllocationOutcome, Allocation } from '@statechannels/nitro-protocol';
 const PROTOCOL = 'virtual-defunding-as-leaf';
