+++ packages/wallet-specs/src/temp-store.ts
@@ -1,0 +1,5 @@
+// TODO: This is temporary until all protocols have been updated
+// to not depend on a global store being defined
+import { Store } from './store';
+const store = new Store();
+export { store };
